requirements.txt
bcrypt==4.1.3
cffi==1.16.0
click==8.1.7
cryptography==42.0.7
Jinja2==3.1.4
MarkupSafe==2.1.5
paramiko==3.4.0
pycparser==2.22
PyNaCl==1.5.0
PyYAML==6.0.1
